+++ components/weapons/index.js
@@ -1,0 +1,5 @@
+// Barrel for the weapons / ballistics layer.
+export { default as Weapon } from "./weapon.js";
+export { default as Bullet } from "./bullet.js";
+export { default as Armor } from "./armor.js";
+export { raycastShape, evaluateImpact, reflect } from "./ballistics.js";